Color rendering

+++ include/view.h
@@ -6,7 +6,7 @@
 // A Cell is a single character drawn to the screen. It contains the character
 // itself, styling information, and additional flags.
 typedef struct Cell {
-    uint32_t fg, bg;
+    uint16_t fg, bg;
     uint16_t flags;
     byte c;
 } Cell;
